requirements.txt
asgiref==3.6.0
black==22.12.0
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
cssselect==1.2.0
dnspython==2.2.1
greenlet==2.0.1
idna==3.4
lxml==4.9.2
mypy-extensions==0.4.3
packaging==22.0
parsel==1.7.0
pathspec==0.10.3
platformdirs==2.6.2
psycopg2==2.9.5
psycopg2-binary==2.9.5
pymongo==4.3.3
python-decouple==3.6
requests==2.28.1
SQLAlchemy==1.4.45
SQLAlchemy-Utils==0.39.0
sqlparse==0.4.3
tomli==2.0.1
urllib3==1.26.13
w3lib==2.1.1
